Spill seems to work correctly though new block types will likely need to be created.

+++ presto-spi/src/main/java/com/facebook/presto/spi/block/array/LongArraySlice.java
@@ -38,7 +38,7 @@ public class LongArraySlice
     @Override
     public long get(int position)
     {
-        return slice.getInt(position * 8);
+        return slice.getLong(position * 8);
     }
 
     @Override
